fix: parser DAO key types follow the widened BCard primary key
BCardId grew from short to int for the 44292-row BCard table, but the
four parsers consuming IDao<BCardDto, TPk> kept short. The startup
registration derives TPk from the entity model, so the mismatch only
surfaced as an unresolvable constructor when the parser was launched.
A contract test now walks every parser constructor and checks each
declared DAO key against the model, turning this class of crash into a
test failure.

+++ src/NosCore.Parser/Parsers/ItemParser.cs
@@ -21,7 +21,7 @@ using System.Threading.Tasks;
 
 namespace NosCore.Parser.Parsers
 {
-    public class ItemParser(IDao<ItemDto, short> itemDao, IDao<BCardDto, short> bCardDao, ILoggerFactory loggerFactory, ILogLanguageLocalizer<LogLanguageKey> logLanguage)
+    public class ItemParser(IDao<ItemDto, short> itemDao, IDao<BCardDto, int> bCardDao, ILoggerFactory loggerFactory, ILogLanguageLocalizer<LogLanguageKey> logLanguage)
     {
         private readonly ILogger<ItemParser> _logger = loggerFactory.CreateLogger<ItemParser>();
         //  VNUM	{VNum}	{Price}
